mv detector logic to detection
