+++ src/lib/detection/mod.rs
@@ -1,3 +1,5 @@
+mod detector;
 mod postprocess;
 
+pub use self::detector::Detector;
 pub use self::postprocess::*;
